Display and intake changes
In this change I made a new subclass for Robot.java. I added a "Sensors" class and as of right now it only has the line sensor but more can be added. We need to finish the intake methods. It should be fairly self-explainatory.

I also added the option to the smartdashboard to display teleop and autonomous data so please check that change.